# tsearch: speed up TSVector.String