--- a/pkg/util/tsearch/tsvector.go
+++ b/pkg/util/tsearch/tsvector.go
@@ -11,10 +11,12 @@
 package tsearch
 
 import (
+	"math/bits"
 	"sort"
 	"strconv"
 	"strings"
 	"unicode"
+	"unicode/utf8"
 
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
@@ -72,6 +74,7 @@ const (
 	weightAny = weightA | weightB | weightC | weightD
 )
 
+// NB: must be kept in sync with stringSize().
 func (w tsWeight) writeString(buf *strings.Builder) {
 	if w&weightStar != 0 {
 		buf.WriteByte('*')
@@ -88,6 +91,14 @@ func (w tsWeight) writeString(buf *strings.Builder) {
 	if w&weightD != 0 {
 		buf.WriteByte('D')
 	}
+}
+
+// stringSize returns the length of the string that corresponds to this
+// tsWeight.
+// NB: must be kept in sync with writeString().
+func (w tsWeight) stringSize() int {
+	// Count the number of bits set in the lowest 5 bits.
+	return bits.OnesCount8(uint8(w & 31))
 }
 
 // TSVectorPGEncoding returns the PG-compatible wire protocol encoding for a
@@ -164,6 +175,7 @@ func newLexemeTerm(lexeme string) (tsTerm, error) {
 	return tsTerm{lexeme: lexeme}, nil
 }
 
+// NB: must be kept in sync with stringSize().
 func (t tsTerm) writeString(buf *strings.Builder) {
 	if t.operator != 0 {
 		switch t.operator {
@@ -217,6 +229,46 @@ func (t tsTerm) writeString(buf *strings.Builder) {
 	}
 }
 
+// stringSize returns the length of the string representation of this tsTerm.
+// NB: must be kept in sync with writeString().
+func (t tsTerm) stringSize() int {
+	if t.operator != 0 {
+		switch t.operator {
+		case and, or, not, lparen, rparen:
+			return 1
+		case followedby:
+			if t.followedN == 1 {
+				return 3 // '<->'
+			}
+			return 2 + len(strconv.Itoa(int(t.followedN))) // fmt.Sprintf("<%d>", t.followedN)
+		}
+	}
+	size := 1 // '\''
+	for _, r := range t.lexeme {
+		if r == '\'' {
+			// Single quotes are escaped as double single quotes inside of a
+			// TSVector.
+			size += 2
+		} else {
+			// Compare as uint32 to correctly handle negative runes.
+			if uint32(r) < utf8.RuneSelf {
+				size++
+			} else {
+				size += utf8.RuneLen(r)
+			}
+		}
+	}
+	size++                   // '\''
+	size += len(t.positions) // ':' or ',' for each position
+	for _, pos := range t.positions {
+		if pos.position > 0 {
+			size += len(strconv.Itoa(int(pos.position)))
+		}
+		size += pos.weight.stringSize()
+	}
+	return size
+}
+
 func (t tsTerm) matchesWeight(targetWeight tsWeight) bool {
 	if targetWeight == weightAny {
 		return true
@@ -247,6 +299,19 @@ func (t TSVector) String() string {
 		term.writeString(&buf)
 	}
 	return buf.String()
+}
+
+// StringSize returns the length of the string that would have been returned on
+// String() call, without actually constructing that string.
+func (t TSVector) StringSize() int {
+	var size int
+	if len(t) > 0 {
+		size = len(t) - 1 // space
+	}
+	for _, term := range t {
+		size += term.stringSize()
+	}
+	return size
 }
 
 // ParseTSVector produces a TSVector from an input string. The input will be