--- a/pkg/util/tsearch/tsvector_test.go
+++ b/pkg/util/tsearch/tsvector_test.go
@@ -187,3 +187,26 @@ func TestParseTSRandom(t *testing.T) {
 		assert.Equal(t, v, v2)
 	}
 }
+
+func BenchmarkTSVector(b *testing.B) {
+	r, _ := randutil.NewTestRand()
+	tsVectors := make([]TSVector, 10000)
+	for i := range tsVectors {
+		tsVectors[i] = RandomTSVector(r)
+	}
+	b.ResetTimer()
+	b.Run("String", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			for _, v := range tsVectors {
+				_ = v.String()
+			}
+		}
+	})
+	b.Run("StringSize", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			for _, v := range tsVectors {
+				_ = len(v.String())
+			}
+		}
+	})
+}