--- a/pkg/util/tsearch/tsvector_test.go
+++ b/pkg/util/tsearch/tsvector_test.go
@@ -188,6 +188,14 @@ func TestParseTSRandom(t *testing.T) {
 	}
 }
 
+func TestTSVectorStringSize(t *testing.T) {
+	r, _ := randutil.NewTestRand()
+	for i := 0; i < 1000; i++ {
+		v := RandomTSVector(r)
+		require.Equal(t, len(v.String()), v.StringSize())
+	}
+}
+
 func BenchmarkTSVector(b *testing.B) {
 	r, _ := randutil.NewTestRand()
 	tsVectors := make([]TSVector, 10000)
@@ -205,7 +213,7 @@ func BenchmarkTSVector(b *testing.B) {
 	b.Run("StringSize", func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
 			for _, v := range tsVectors {
-				_ = len(v.String())
+				_ = v.StringSize()
 			}
 		}
 	})